fix: consistent file-not-found errors across edit, send_file, and read
edit's `realpath` and send_file's inline stat catch leaked raw Node ENOENT
messages (`lstat error`, `42: ... statx`) instead of actionable guidance.

Promote read's missing-file renderer with its "did you mean" sibling
suggestions into `FsErrors.statOrThrow`, then route edit, send_file, and
read through the shared helper. glob and grep already use it and now pick
up the suggestions for free.
